more todo cleanups
